+++ bld/wv/h/filelcl.h
@@ -31,8 +31,8 @@
 
 extern error_idx        LocalErase( char const * );
 extern void             LocalErrMsg( sys_error, char * );
-extern unsigned         LocalRead( sys_handle, void *, unsigned );
-extern unsigned         LocalWrite( sys_handle, const void *, unsigned );
+extern size_t           LocalRead( sys_handle, void *, size_t );
+extern size_t           LocalWrite( sys_handle, const void *, size_t );
 extern unsigned long    LocalSeek( sys_handle, unsigned long, seek_method );
 extern sys_handle       LocalOpen( char const *, open_access );
 extern error_idx        LocalClose( sys_handle );
